+++ src/atoms/rec.ts
@@ -1,7 +1,7 @@
 import { atom } from "recoil";
 import { RecWriteReqType } from "types/rec.type";
 
-export const FeedTitleAtom = atom<RecWriteReqType>({
+export const RecWriteAtom = atom<RecWriteReqType>({
   key: "recWrite",
   default: {
     title: "",
