+++ src/components/common/Header/index.tsx
@@ -1,3 +1,4 @@
+
 import * as _ from "./style";
 import { useSetRecoilState } from "recoil";
 import { signInModalAtom, signupModalAtom } from "atoms/header";
