+++ src/components/pages/common/Feed/ArticleComment/type.ts
@@ -1,5 +1,4 @@
 export interface ArticleCommentProps {
   writer: string;
   content: string;
-  is_mine?: boolean;
 }
